Optimize the repo code analysis.
Since lizard might fail to analyze a code file and make the memory boom,
the new approach get all source files and analyze them one by one, and
calculate the metrics manually.

The single file analysis function is decorated with a timeout, if lizard
failed to analyze within the timeout, we just skip it(assume that lizard
cannot handle it).